book, journal, timetable change

+++ application/views/admin/book/getall.php
@@ -53,54 +53,55 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
                                 <tbody>
                                     <?php
                                     $count = 1;
-                        if(!empty($listbook)){
+                                    if (!empty($listbook)) {
 
-                    
-                                    foreach ($listbook as $book) {
-                                       
-                                        ?>
-                                        <tr>
-                                            <td class="mailbox-name">
-                                                <a href="#" data-toggle="popover" class="detail_popover"><?php echo $book['book_title'] ?></a>
-                                                <div class="fee_detail_popover" style="display: none">
-                                                    <?php
-                                                    if ($book['description'] == "") {
-                                                        ?>
-                                                        <p class="text text-danger"><?php echo $this->lang->line('no_description'); ?></p>
+
+                                        foreach ($listbook as $book) {
+
+                                            ?>
+                                            <tr>
+                                                <td class="mailbox-name">
+                                                    <a href="#" data-toggle="popover" class="detail_popover"><?php echo $book['book_title'] ?></a>
+                                                    <div class="fee_detail_popover" style="display: none">
+                                                        <?php
+                                                        if ($book['description'] == "") {
+                                                            ?>
+                                                            <p class="text text-danger"><?php echo $this->lang->line('no_description'); ?></p>
                                                         <?php
                                                     } else {
                                                         ?>
-                                                        <p class="text text-info"><?php echo $book['description']; ?></p>
+                                                            <p class="text text-info"><?php echo $book['description']; ?></p>
                                                         <?php
                                                     }
                                                     ?>
-                                                </div>
-                                            </td>
-                                            <td class="mailbox-name"> <?php echo $book['book_no'] ?></td>
-                                            <td class="mailbox-name"> <?php echo $book['isbn_no'] ?></td>
-                                            <td class="mailbox-name"> <?php echo $book['publish'] ?></td>
-                                            <td class="mailbox-name"> <?php echo $book['author'] ?></td>
-                                            <td class="mailbox-name"><?php echo $book['subject'] ?></td>
-                                            <td class="mailbox-name"><?php echo $book['rack_no'] ?></td>
-                                            <td class="mailbox-name"> <?php echo $book['qty'] ?></td>
-                                            <td class="mailbox-name"> <?php echo $book['qty']-$book['total_issue'] ?></td>
-                                            <td class="mailbox-name"> <?php echo ($currency_symbol . $book['perunitcost']); ?></td>
-                                            <td class="mailbox-name"> <?php echo date($this->customlib->getSchoolDateFormat(), $this->customlib->dateyyyymmddTodateformat($book['postdate'])); ?></td>
-                                            <td class="mailbox-date no-print text text-right">
-                                                <?php if ($this->rbac->hasPrivilege('books', 'can_edit')) { ?> 
-                                                    <a href="<?php echo base_url(); ?>admin/book/edit/<?php echo $book['id'] ?>" class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
-                                                        <i class="fa fa-pencil"></i>
-                                                    </a>
-                                                <?php }if ($this->rbac->hasPrivilege('books', 'can_delete')) { ?> 
-                                                    <a href="<?php echo base_url(); ?>admin/book/delete/<?php echo $book['id'] ?>"class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
-                                                        <i class="fa fa-remove"></i>
-                                                    </a>
-                                                <?php } ?>
-                                            </td>
-                                        </tr>
-                                        <?php
-                                        $count++;
-                                    }
+                                                    </div>
+                                                </td>
+                                                <td class="mailbox-name"> <?php echo $book['book_no'] ?></td>
+                                                <td class="mailbox-name"> <?php echo $book['isbn_no'] ?></td>
+                                                <td class="mailbox-name"> <?php echo $book['publisher'] ?></td>
+                                                <td class="mailbox-name"> <?php echo $book['author'] ?></td>
+                                                <td class="mailbox-name"><?php echo $book['subject'] ?></td>
+                                                <td class="mailbox-name"><?php echo $book['rack_no'] ?></td>
+                                                <td class="mailbox-name"> <?php echo $book['qty'] ?></td>
+                                                <td class="mailbox-name"> <?php echo $book['qty'] - $book['total_issue'] ?></td>
+                                                <td class="mailbox-name"> <?php echo ($currency_symbol . $book['bookprice']); ?></td>
+                                                <td class="mailbox-name"> <?php echo date($this->customlib->getSchoolDateFormat(), $this->customlib->dateyyyymmddTodateformat($book['postdate'])); ?></td>
+                                                <td class="mailbox-date no-print text text-right">
+                                                    <?php if ($this->rbac->hasPrivilege('books', 'can_edit')) { ?>
+                                                        <a href="<?php echo base_url(); ?>admin/book/edit/<?php echo $book['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
+                                                            <i class="fa fa-pencil"></i>
+                                                        </a>
+                                                    <?php }
+                                                if ($this->rbac->hasPrivilege('books', 'can_delete')) { ?>
+                                                        <a href="<?php echo base_url(); ?>admin/book/delete/<?php echo $book['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
+                                                            <i class="fa fa-remove"></i>
+                                                        </a>
+                                                    <?php } ?>
+                                                </td>
+                                            </tr>
+                                            <?php
+                                            $count++;
+                                        }
                                     }
                                     ?>
                                 </tbody>
@@ -115,7 +116,8 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
                         </div>
                     </div>
                 </div>
-            </div><!--/.col (left) -->
+            </div>
+            <!--/.col (left) -->
             <!-- right column -->
         </div>
         <div class="row">
@@ -124,19 +126,20 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
             <div class="col-md-12">
                 <!-- Horizontal Form -->
                 <!-- general form elements disabled -->
-            </div><!--/.col (right) -->
-        </div>   <!-- /.row -->
+            </div>
+            <!--/.col (right) -->
+        </div> <!-- /.row -->
     </section><!-- /.content -->
 </div><!-- /.content-wrapper -->
 <script type="text/javascript">
-    $(document).ready(function () {
+    $(document).ready(function() {
         var date_format = '<?php echo $result = strtr($this->customlib->getSchoolDateFormat(), ['d' => 'dd', 'm' => 'mm', 'Y' => 'yyyy',]) ?>';
         $('#postdate').datepicker({
             // format: "dd-mm-yyyy",
             format: date_format,
             autoclose: true
         });
-        $("#btnreset").click(function () {
+        $("#btnreset").click(function() {
             /* Single line Reset function executes on click of Reset Button */
             $("#form1")[0].reset();
         });
@@ -148,12 +151,15 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
 
 <script type="text/javascript">
     var base_url = '<?php echo base_url() ?>';
-    function Popup(data)
-    {
+
+    function Popup(data) {
 
         var frame1 = $('<iframe />');
         frame1[0].name = "frame1";
-        frame1.css({"position": "absolute", "top": "-1000000px"});
+        frame1.css({
+            "position": "absolute",
+            "top": "-1000000px"
+        });
         $("body").append(frame1);
         var frameDoc = frame1[0].contentWindow ? frame1[0].contentWindow : frame1[0].contentDocument.document ? frame1[0].contentDocument.document : frame1[0].contentDocument;
         frameDoc.document.open();
@@ -179,7 +185,7 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
         frameDoc.document.write('</body>');
         frameDoc.document.write('</html>');
         frameDoc.document.close();
-        setTimeout(function () {
+        setTimeout(function() {
             window.frames["frame1"].focus();
             window.frames["frame1"].print();
             frame1.remove();
@@ -190,18 +196,18 @@ $currency_symbol = $this->customlib->getSchoolCurrencyFormat();
     }
 
 
-    $("#print_div").click(function () {
+    $("#print_div").click(function() {
         Popup($('#bklist').html());
     });
 
 
-    $(document).ready(function () {
+    $(document).ready(function() {
         $('.detail_popover').popover({
             placement: 'right',
             trigger: 'hover',
             container: 'body',
             html: true,
-            content: function () {
+            content: function() {
                 return $(this).closest('td').find('.fee_detail_popover').html();
             }
         });
